Phase 1 – Mechanical Extraction is fully executed.
Summary of delegated subtasks and outcomes:

1. GUI tier split
   • Created font_utils.{h,cpp}, theme_utils.{h,cpp}, event_dispatch.{h,cpp}.
   • Trimmed gui_interface/gui_interface.cpp to 315 LOC; updated includes and CMake.
   • Build & smoke-test succeeded.

2. Main entry split
   • Added main_gui_core.cpp (179 LOC), main_gui_views.{h,cpp} (372 LOC).
   • main_gui.cpp reduced to empty shim; CMake updated.
   • Unit test test_chronological_layout passes; GUI runs.

3. Renderer extraction
   • Added graph_renderer_core.cpp (449 LOC), graph_drawing_utils.{h,cpp} (179 LOC), camera_utils.{h,cpp} (42 LOC).
   • graph_renderer.cpp trimmed to 14 LOC; theme helpers moved to theme_utils.
   • CMake adjusted; Debug & Release builds clean; visual parity confirmed.

4. Layout extraction
   • Added force_directed_layout.{h,cpp}, graph_layout_recursive.{h,cpp}, spatial_hash.{h,cpp}.
   • graph_layout.cpp now façade ≤ 500 LOC.
   • tests/force_convergence_test passes; GUI layout behaviour unchanged.

5. Persistence extraction
   • Added sqlite_connection.{h,cpp}, message_store.{h,cpp}, model_store.{h,cpp}, settings_store.{h,cpp}, database_fwd.h.
   • database.cpp slimmed to 91 LOC; new static library “persistence” linked in CMake.
   • All unit & integration tests green; GUI data persists correctly.

Across all subtasks every new/edited source file is ≤ 500 LOC, project builds in Debug & Release with zero warnings, and all automated & manual verifications succeed.

Phase 1 refactor is complete and compliant with docs/phase1.md objectives.
