+++ database_fwd.h
@@ -1,0 +1,6 @@
+#pragma once
+
+class SQLiteConnection;
+struct MessageStore;
+struct ModelStore;
+struct SettingsStore;
