+++ graph_layout_recursive.h
@@ -1,0 +1,18 @@
+#ifndef GRAPH_LAYOUT_RECURSIVE_H
+#define GRAPH_LAYOUT_RECURSIVE_H
+
+#include <map>
+#include "extern/imgui/imgui.h"
+#include "gui_interface/graph_types.h"
+
+void CalculateNodePositionsRecursive(
+    GraphNode* node, 
+    ImVec2 current_pos,
+    float x_spacing, 
+    float y_spacing, 
+    int depth, 
+    std::map<int, float>& level_x_offset, 
+    const ImVec2& canvas_start_pos
+);
+
+#endif // GRAPH_LAYOUT_RECURSIVE_H
