+++ tools_impl/deep_research_tool.h
@@ -4,4 +4,4 @@
 class ChatClient;
 class UserInterface; // Forward declaration
 
-std::string perform_deep_research(PersistenceManager& db, ChatClient& client, UserInterface& ui, const std::string& goal);
+std::string perform_deep_research(Database& db, ChatClient& client, UserInterface& ui, const std::string& goal);
