+++ tools_impl/read_history_tool.h
@@ -2,4 +2,4 @@
 #include <string>
 #include "database.h"
 
-std::string read_history(PersistenceManager& db, const std::string& start_time, const std::string& end_time, size_t limit);
+std::string read_history(Database& db, const std::string& start_time, const std::string& end_time, size_t limit);
