+++ tools_impl/web_research_tool.h
@@ -4,4 +4,4 @@
 class ChatClient;
 class UserInterface; // Forward declaration
 
-std::string perform_web_research(PersistenceManager& db, ChatClient& client, UserInterface& ui, const std::string& topic);
+std::string perform_web_research(Database& db, ChatClient& client, UserInterface& ui, const std::string& topic);
